Rust lexer has to be able to handle singular quotes '/"
In a valid Rust programm quotes occur in pairs to form character,
string or byte string literals. However while inputting the programm
the lexer must still be able to handle them. The lexer has to produce
a token in that case so that the NetBeans infrastructure relying on the
token stream can work.

It is not a problem, that the resulting token stream can't be parsed.

Concrete cases:

Removing parts of single quote structure. The pipe symbols are not
part of the code. The area from the first pipe to the second pipe is
selected and removed.

fn main() {
    println!('|x');
}
|

Enter single quote ('). The pipe symbol is not part of the code, at
that position a single quote is entered.

fn main() {
    println!(|

This partially reverts b62889b5c2799cd58c925e8fd560088d34947418 and
introduces a different fix, that handles the "lonely" characters as
explicit tokens.