Add support for Wildfly 28.0.0+
NetBeans communicates to Wildfly via management api. This functionality depends
on several Wildfly classes which are pulled from the Wildfly installation.

It seems that in 28.0.0 some classes and jars were moved which prevents NetBeans
from communicating to Wildfly.

This change simplify the current logic gathering different jars and just use
`jboss-cli-client` which should contain all necessary classes which are needed
to connect to Wildflys management interface.