Fix unstable unittest org.netbeans.modules.maven.ModuleInfoSupportTest
25347dfe7f75e540d751f06b541320dc6c5511ce introduces a new unittest,
that proved to be unstable in CI/CD pipeline. The failure can be
reproduced after several tries. Adding 

workDir.getFileSystem().refresh(true);

seems to fix the issue. Hypothesis:

The new module files are written outside the filesystem infrastructure
of NetBeans and the cache of the javac integration does not pick up the
module files created by the javac-Tool.

Apart from the fix, style of the code was improved.