Bugfix: Bring back singleton option when creating a new `Session Bean` and interfaceless EJBs
The options shown when selecting `New > Session Bean` are based on the capabilities of the used enterprise project.

The check missed some Jakarta EE versions and so were disabled.

As EJB and CDI specs are backwards compatible the code was refactored so that only a minimal version and the profile type is checked to match that EJB or CDI version.

This fixes #4892 and fixes #4830