Fix: Committing SQL files does not change blue color on the file tab in editor
The construction of the editor panel for SQL is a bit special to
support both using it as SQL console and file editor and allow it to
execute sql statements.

In normal editors multiview elements are initialized by the
infrastructure and #initializeCloneableEditor is called by it to
connect the DataNode and the editor. This was missing for the SQLEditor.

Closes: #6224