Terminal input ignores detected encoding
It was observed that on a german windows installation with JDK 17 the
terminal correctly shows german umlauts, but fails on input.

The terminal subsystem has code to detect the charset used in the
terminal.

Investigation then leads to StreamTerm, where the InputStream for data
coming from the terminal is adapted with the detected charset, while
the user input is ran through an OutputStreamWriter without charset
set. Therefore the user input is encoded with Cp1250 (default on
german windows setups). This leads to garbled input.

The fix applies the same logik, that is already applied to the
InputStream, also to the OutputStream.