Switch graalvm CI setup to the graalvm/setup-graalvm action.
GraalVM has a new versioning scheme and the URLs changed too, lets
use the opportunity to switch to the setup-graalvm action.

This will also move the minimal version to JDK 11 since graalvm CE
has only support for 17 and 20. 11 appears to be already EOL.

The R component seems to be no longer available.