Fix Project Group window EDT race condition
The modal project-group dialog can open a modal create-group dialog
as last action before it disposes. This apparently creates a race
condition since the create-group dialog can receive a parent which just
disappeared - leading to weird UI behavior.

SwingUtilities.invokeLater seems to be sufficient to make sure the
first dialog is gone before the second opens so that it can get a
proper parent, which is usually the main window.

+ minor cleanup