Improve the code completion for enum cases
There is an error in the following example because `case B = self::` is not a complete statement.
Thus, the result of parsing doesn't recognize the enum declaration.
This means that some features don't work correctly.
(e.g. The enum declaration doesn't exist in the navigator. The code completion doesn't work.)

```php
enum Example: int {
    case A = 1;
    case B = self::
}
```

To improve this, add the `ASTErrorExpression` as a result of an enum case initializer part.

- Fix the parser
- Add unit tests