Drop special mimetype for karma test configuration
The karma test configuration (default filename: karma.conf.js) was 
mapped to the special mimetype text/karmaconf+javascript.

The mimeresolver for that matched for all files, that ended with js and
contained "karma" or "conf". This collides with the JS editor
infrastructure, that does not allow for this subclassing. Oen such
problem was, that for file JS folding was provided, but the fold types
were not provided, raising exceptions from GsfFoldManager.

Instead of special casing the file to a mimetype, the action is now
tied to all JS files, but only activated (and visible) on karma*.conf.js
files.

Closes: #6879