[GITHUB-6970] MultiSourceRootProvider: reduce logging verbosity
The ClassPathProvider is also invoked for files that are only temporary
present. The files are only present for a short time and between the
FileObject#isValid call and the FileObject#asBytes call the file is
removed leading to an IOException.

This situation is a race condition and happens often enough for users
to be annoyed. As this is common and the scanning error is not fatal
for normal operation, this should normally not be reported.

Closes: #6970