Fix empty Test Results with maven surefire plugin.
When running junit tests with maven and the surefire plugin 3.0.0+, the log
level can be included in the output for the plugin.  This output is used to
detect which class is being tested for display in the Test Results tab.
Updated the class matching regex to ignore the log level if present.