Handle JAR-URL as codesource for javascript editor module
The fallback code in ClassPathProviderImpl#getJsStubs should only ever
be called in unittests, but also seems to be sometimes called in
production. In that case the URL reported by
Class#getProtectionDomain#getCodeSource#getLocation is a JAR-URL. That
URL can't be directly mapped to a local file. The file URL has to be
extracted first.

Closes: #7157