Bugfix entity generation in plain java (non webapp) projects
For plain java projects there is no JPADataSourcePopulator available,
so it must not be called to populate the schema list. The schema
combobox for this case should then not be shown.

For the case when no JPADataSourcePopulator and no schema file is
present, the default selected datasource selector should not be
server datasources, but local data sources.

Closes: #7144