JS: Provide model data for meta properties new.target and import.meta
Both new.target and import.meta are declared as special properties with
special semantic as they are properties on keywords. The approach taken
here is to declare a helper object "metaproperties", declare the "new"
and "import" properties and their children there. After parsing the
properties are relocated from metaproperties into the global object,
where they will be picked up by the JsCompletion.