XML Catalog GF/Payara: Ensure only existing schemas are resolved
Before this change the InputSource for some of the schema referenced by
the JSF files failed to load when GF 5.X was registered. The problem
was, that the Catalog code resolved jakartaee_web_services_client_2_0.xsd
to a non existing local path and in turn the XML parser failed to parse
the JSF component libraries.

Closes: #7137