JSF: Support composite components in resource folder defined by jakarta.faces.WEBAPP_RESOURCES_DIRECTORY
Together with the packages and namespaces the JavaEE -> JakartaEE
transition also invalidated/renewed the names of context parameters.

This also affects the resources folder which is defined by the context
parameter javax.faces.WEBAPP_RESOURCES_DIRECTORY or
jakarta.faces.WEBAPP_RESOURCES_DIRECTORY.

Both parameters need to be checked if they are present and are now
considered to search for composite components (the last defined
definitions wins).

Closes: #7226