Drop GraalVM test runs from CI/CD
The GraalVM tests have tested outdated GraalVM execution modes. In the
beginning GraalVM had the mode where the base VM was installed and
additional languages were installed in to that distribution.

That mode was replaced by either distinctive distributions for the
supported languages or by loading the JARs into the classpath.

The latter approach is in line with the NetBeans approach to loading
JS support through GraalVM/Truffle.