Fix usage of ToggleBookmarkAction in split windows scenario
It was observed, that the "Goto Bookmark" actions cause the goto action
in the correct window (the focused one or the one the toolbar is
located in), while the "Toogle Bookmark" action invokes the
action for the first editor window.

The implementation of the ToggleBookmarkAction implies that the button
should reflect whether or not there s a bookmark on the current line
(rendering the button active/pressed when there is a bookmark). It was
observered, that this does not work.

So instead of fixing the complex implementation, the implementation for
ToggleBookmarkAction was aligned with the "GotoBookmarkAction",
inheriting the "find the right text component" behavior from BaseAction.

Closes: #4155