After 562cfade954ea27308de01738813d050d1b57ce1 ContextAction.Performer must be context aware
With 562cfade954ea27308de01738813d050d1b57ce1 state tracking was
introduced to ContextAction. Both ContextAction and Performer are now
tied to the lookup and thus a new Performer must be used when
createContextAwareInstance is invoked.

To support this instead of a performer, a performer factory is
supplied to the ContextAction and the Performer is created in the
constructor of the ContextAction.