Popupmenu for Terminals should be invoked for target tab, not for visible tab
The NB Terminals can open multiple tabs. Before this change the popup
menu opened in the "Tab"-Area shows actions, that are invoked on the
visible tab, even when the mouse cursor hovers over another tab.

With this change, if the cursor hovers of a terminal tab, that tab
is the target for the action.