Fix platform/openide.util.ui test on JDK 11+
Remove the testSignificantWhitespace test as it is broken since JDK9 and
tests something clearly not supported by the documentation of
XMLUtil#write:

> The precise output format is not guaranteed but this method will
> attempt to indent it sensibly.

So reindention must be expected. For an XML document without a DTD/XSD
all whitespace between and in elements must be considered significant.
If this behavior is really needed, a new codepath, which preseves
whitespace, would be needed.

The XMLUtilTest is renamed to XMLUtil2Test.java so it does not clash
with the identically named test in `openide.util`.