Remove hard dependency on presence of ProjectManageImplementation in platform/openide.nodes
The ProjectManagerDeadlockDetector tries to get the project mutex by
reflectively instantiating the org.netbeans.api.project.ProjectManager.
In the constructor a ProjectManageImplementation is looked up from the
default Lookup and if not found, raises an IllegalStateException.

As the construction of the ProjectManager is done as part of the class
initialization (static field "DEFAULT"), the JVM wraps the exception
into an ExceptionInInitializerError. That also needs to be caught.