JSF Templates need to take into account new namespace schema for Jakarta Faces 4+
In the past the core namespaces for JSF were named:

http://xmlns.jcp.org/jsf/<core|facelets|html>

And in templates "http://xmlns.jcp.org" was pulled out as $nsLocation.
(JSF namespace changed from 2.1 to 2.2).

With JSF 4 another namespace change was enforced as part of the
donation. The namespaces are now named:

jakarta.faces.<core|facelets|html>

This needs to be considered by templates.