Fix java completion sort for not-imported items.
LazySortText implements CharSequence but missed a toString() method.

It would use Object#toString() and since it is used as
comparator input, it lead to undesired results.

This was the case with 'List.of' for example. ('List' must not have an
import yet)

This commit introduces utility methods which allow the linking
of CharSequences like LazySortText without toString() concatenation.