Editor: Enable display of characters unsupported by chosen editor font
It was noticed, that in some contexts Swing was able to display
characters unsupported by the currently chosen font. The problem was
observed with emojis. Investigation showed, that the problem appears if
a font is explicitly constructed via the constructor, but vanishes if
the font is accessed via StyleContext#getFont.

The difference is that StyleContext#getFont returns a CompositeFont,
that fallsback to other fonts for undefined glyphs.

As CompositeFont is located in the sun.font package, it can't be
accessed directly. Instead the font acccess via the DefaultStyleContext
is used.