LSP client stops working on file after renaming it
The LSP protocol requires that when a file is renamed the client
signals this to the server by sending a didClose for the old location
and after that a didOpen with the new location.

The change assumes, that the IDE always goes through the DataObject to
issue the rename/move. Changes through the raw FileObject are not
tracked.

Closes: #7806