Fix parsing of @keyframes rules
Before 23ed11129c5eee8d6806af906fbac9c1b5c34daf "CSS: Improve handling
of generic-at-rules (for example used by tailwind)" the CSS parser
returned the body of generic @-rules as a plain list of tokens. That
changed and the @keyframes rule needs to be special cased. There was
already a special case for the @-webkit-keyframes variant, which is
identical to the final variant and differs only in the introducing
rule name.

The vendored variants @-moz-keyframes and @-o-keyframes are also
supported.

Closes: #8052