On MacOS, add Command+O as a shortcut for Open (default action) in TreeView (Projects, Files tabs etc.), like in Finder.
Note that that the same shortcut is also used by default for the Go to Type action. But the action to open a TreeView node only takes precedence when the TreeView (e.g. Projects tab) is actually in focus, which makes this conflict less of a problem for the IDE.