added record in SAMPLE_Spaces
to show the effect of propery spaceswithinMethodDeclaration on record.
See issue #7043.

moved Sample spaces from bundle to text block.

To show effect of #7043.

added test RecordFormattingTest to show problem  \#7043

There are two tests in the test class, one that passws with the spaceWithinMethodDeclParens
set to false, and one that fails when this is set to true.