On FlatLAF, add custom icons for file choosers and OptionPane to match the style of other NetBeans icons.
The icons for OptionPane are 32x32 pixel version of the info/error/warning/question icons (scaled up from similar 16x16 SVG icons, but maintaining the 1px border width as this looks better).

The icons for file chooser are "folder", "file", "parent folder", and "home folder".

To keep the SVG loader module optional for NetBeans Platform apps, PNG versions of the icons are used here. SVG versions will be added in a separate commit (together with a bunch of other SVG files).