Add another batch of SVG icons, drawn by myself. This is a continuation of the effort to make NetBeans look good on HiDPI/Retina screens.
There are 107 new icons in total, copied into many more places where duplicate bitmap icons were identified. Some existing icons were also adjusted, and some existing icons were copied into new locations where additional duplicates were identified.

I prioritized icons that are commonly seen in the IDE, and which happen to be easy to draw, as well as icons that are similar to already-drawn ones or simple variations or combinations thereof.

The changes in this commit consist purely of copying SVG files into various locations in the repo; this was done by the hidpi-icons script at https://github.com/apache/netbeans-tools/tree/master/icon-scripts/hidpi-icons . I will open a separate Pull Request for that repo, to update the Adobe Illustrator file and icon mapping files.

The 'icons.html' file that is generated by the hidpi-icons script is now included as part of the commit, in the root of the NetBeans repo. It includes a preamble with useful information and links to the README.