csl.api: Provide dummy/empty implementation for language services, that might not be present
There is a mismatch in the CSL API/SPI: The API declared by the
Language class requires the methods:

- getOccurrencesFinder
- getSemanticAnalyzer
- getIndexSearcher
- getOverridingMethods

to return non-null values. This is a problem as the implementation
relies on the corresponding methods in DefaultLanguageConfig to provide
the implementations. These methods are declared to be allowed to return
null.

This mismatch has to be fixed in Language, there the "null" situation
is detected and a no-op implementation is returned.