improved formatting of RefactoringTestBase.
Add expected values for passing test in InnerOuterRecordTest.
The existing javaformatting did not conform to any standard.
To make it possible to work with that file, default netbeans formatting
has been applied. That leads to some white space changes only, so be it.

Further added method assertLinesEqual to get better readable output when
comparing expected and acyla multi line string values, as in comparing
or diffing files.

Part of effort to find the problem described in #7044