add source levels up to java 17, make reindex optional
since Netbeans requires 17 to build and run, it is reasonable to expect
that in source levelutils too.

In RefactoringTestBase the repo is always reindexed, even when not
required.
As per documentation this is an expensive operation and taxes your
patience.

remove modifiers in variable member, so the do not show up

Modifiers (private final in particular) show up in record header,
making the resulting source code illegal and different from expected.
The removal is done by rewriting the member tree where variables occur.

make sure static fields of inner records survive field removal.

ensure static fields remain in inner record.

normal fields are modified by removing private final, but this should
not happen
to static fields.
This is part of making inner records work.

avoid use of raw types.

pass declared type further down.

Original code always passes down CLASS, now RECORD is passed down as
such. It should not make a difference for class, interface, or enum.

spelling

RECORD should be equivalent to class, interface and enum.

Add RECORD_COMPONENT as allowed element types

add isInnerRecord boolean field. Some reformatting.

For the reformatting: sorry. Stick to the formatting rules.

For the field: Record does not fit the standard class definition mold
because of the record header which has parenthesis before the first curly brace
 and the compact constructor that needs special attention.
Somehow the information that ithe inner tis a inner record must be forwardable to
the processors.

The author comment is modified by some part.

The @author line is substituted by some yet unknown processor,
so exclude it from the result comparison to make the tests less britle
because of dependence of the user property.

generateOuter sounds like a boolean, but it is the name of the new inner

match expected to actual after formatting.

This test passes, showing that the Compact constructor in the outer
survives the refactoring.

formatting in expected outer

cleanup. Permanently remove deprecated boolean cancelRequest

cancelRequest is not only deprecated but has no use at all and was
already replaced (function wide) by and atomicBoolean cancelRequested

This commit cleans up the tests, may add a print that needs removing.
For now: I am able to refactor inner record to outer but are still
lacking the understandig of how to modify a MethodTree or the equivalent
source code, and how to write a compact constructor instead.

remove deprecated boolean

no need to handle method special here.

make test helper static to make it usable in more places.