Prefer Icon.paintIcon over Graphics.drawImage to preserve HiDPI icon resolution.
Various SVG icons were previously shown in a lower resolution due to being painted with Graphics.drawImage(ImageIcon, x, y, null) instead of Icon.paintIcon(null, Graphics, x, y). This PR changes the former to the latter in all the simple cases that were found throughout the codebase (by grepping for 'drawImage' and manually reviewing each case).

For example, the lightbulb icons in the editor gutter, and the actions in the menu that shows up when you click them, were shown in low resolution on Windows 11 with 150% HiDPI scaling enabled.

This is mostly an improvement for screens with a fractional HiDPI scaling (e.g. 150%) on Windows. On 200% scaling (including on MacOS), icons are drawn properly even with drawImage, due to ImageUtilities returning MultiResolutionImage instances nowadays with 1x and 2x resolutions present.