Add javaagent to patch WClipboard
It was observed, that clipboard on windows sometimes reacts flaky.
Investigation hints that missing synchronization in the AWT
implementation/integration. Testers reported, that the changes
implemented here:

 - the change handling from the system clipboard is dispatched into the
   event loop and
 - the change handling is modified to be called synchronized

It is expected, that fixing this upstream in the JDK will take longer
and so runtime patching must be considered.

Runtime patching in this case is accomplished using the javaagent
mechanism. The IDE launcher is modified to load an agent, that will
intercept loading classes and modify WClipboard while it is being
loaded.

The transformer is written, so that it will only transform the target
class and will act transparantly for all other classes.