finally solves issue #7044.
This is hopefully my final attempt.

This PR solves issue #7044

The issue describes that record info was lost when doing and inner to outer or pull up refactoring.

In this pull request

tests to show the issue.
junit test helper to make test output better understandable: AssertLinesEqualHelpers.java
added RECORD case label where appropriate, for instance where the record template should be used.
added factory for Record classTree in TreeFactory and TreeMaker
added/improved methods in CasualDiff and VeryPretty to pick op parameters from record in constructor, so that
var args are preserved.