update: update TextMark/Markdown tooling
desc: The TextMark JSON description is very much out of date (i.e. years). This change brings in the latest version of the JSON description.

In addition, update the TextMark plugin to 0.14.1 version.

testing: built with no issues. Loaded a README.md file into the IDE and it previewed correctly.