Improve support for using latest Gradle version in new projects.
Adds a new method to GradleDistributionManager to query the current (latest)
version from the Gradle web service.

Adds support for passing a Gradle version into TemplateOperation.InitOperation
that is passed in to GradleConnector.useGradleVersion

Passes the current version in to the init step in the wizard.

Replaces sometimes temperamental support added in #7029