prevent NPE by checking return value befor use
While working on an unrelated subject I got a error notification that
mentiones an NPE in the modified file

simply add a check of the retrun value before dereferencing it in the
following for-each statement. The result is that no elements will be
added to the list in the for-each loop, which looks okay to me in that
case.