Disabling "Keep Marks" highlighting setting has no effect for PHP
NetBeans allows users to configure whether to keep the occurence
hightlighting when the cursor is moved off the origin symbol. If
"Keep Marks" is enabled the highlights are kept until another symbol
is marked by the caret. If it is disabled the highlights are removed
once the caret is not on the symbol anymore.

There are two problems:

- the CSL integration of OccurrenceFinder ignores updates with empty
  highlights (essentially "Keep Marks" was always on for CSL languages)
- the marks are cleared always and thus without the problem in the CSL
  integration would act as if "Keep Marks" is always off