Maven JUnit: Stabilize unit test result extraction
It was observed, that also for singlethreaded runs the XML reports are
sometimes created delayed. The existing retry logic for parallel runs
needs to be enhanced to cover the single threaded case and also more
than one retry.

Observed with a trivial nested testcases at least one retry was
observed.