SourceUtils#classNameFor does not handle nested classes correctly
The assumption in the method is, that the names of nested classes are
placed into an index, but that is not the case. In the index only
toplevel classes are recorded. So this:

```
package test;

public class Demo {
static class Nested {}
}

class NonPublic {}
```

will yield:

```
test.NonPublic
test.Demo
```

But the nested class `test.Demo.Nested` will not be found! So use the
index only as a validator if a NestedClass is provided.