lsp.client: Callers of LSPBindings#getBindingsImpl extect result to be non null
Files whose name starts with vcs- and lie inside the java.io.tmpdir
will yield `null` for the call to LSPBindings#getBindingsImpl. This
is not expected by callers and breaks the contract of that method
that declares it to yield non-null values.

This change turns the null into an empty list.