GSF/CSL Languages: Centralize "Enable Occurrences Hightlight" and "Keep Marks" Handling
According to the documentation of OccurrencesFinder the occurrences
scanning should be stateless. The current API holds a calculate and
a getter method, but in the spirit of keeping it mostly stateless,
more state should not be added.

The OccurrencesFinder implementation returns a set of occurrenes
corresponding to the currently highlighted area. These areas are
described by integer arrays (2 elements, start and end index). Given
the return types these areas are only valid until the next edit
occurrs.

On each input the highlights are recalculated. That works fine until
"Keep Marks" is activated. "Keep Marks" keeps the marks, when the caret
is moved of a highlightable variable. If "Keep Marks" is enabled the
marks will be retained then until the caret is placed onto another
highlightable element. The marks in this case must adapt to changes of
the content of the element. Swing Document provides the "Position"
class for this. That class is used by GsfSemanticLayer to hold the
areas returned by the OccurrencesFinder implementation. So to support
"Keep Marks" the CSL/GSF layer has to be aware of this, so that the
GsfSemanticLayer can retain the highlights.

In addition to this there is no need to scan for occurrences if the
highlighting is disabled, so that also moves to the CSL/GSF layer.

Closes: #8803