Improve goto search functionality
 - make exact search mode follow the case sensitivity settings
 - add missing exact search instructions to file search label (other
   windows had those)
 - file search: selection should stick to first item until user changes
   selection. Async search could otherwise insert better matches above
   the selected item
 - disable "prefer opened projects" by default
 - CI: enabled tests
 - other minor improvements

the "prefer opened projects" setting can lead to confusing results
when good matches are displayed far away from the first (worse) match,
potentially outside of the visible viewport. The implementation
itself works correctly but it might be better to change it into
a filter instead of model comparator in future and consolidate
the UI/settings of the 3 search dialogs.