Fixes most layer position warnings and improved logging
The warning message will now include the position of the services
to make picking a value easier.

Tested activation of all clusters and maven project creation, only one
root '/' warning is left to fix in the log.

Might make the layer behave more deterministic in some cases.

conflict fixes:

same position 300 for both
UI/ToolActions/Files/org-netbeans-modules-diff-tree-RecursiveDiffAction.shadow and
UI/ToolActions/Files/org-netbeans-modules-favorites-Add.shadow
 -> RecursiveDiffAction = 250

same position 650 for both Templates/Project/GradleGroovy and Templates/Project/Native
Native -> 660

ordering fixes:

Not all children in Windows2/Modes/output/ marked with the position attribute:
[TextToolsTopComponent.wstcref], but some are:
[watchesView.wstcref, NotificationCenterTopComponent.wstcref, ...]
 -> TextToolsTopComponent = 3400

Not all children in OptionsDialog/Editor/ marked with the position attribute:
[InlineHints], but some are:
[CodeCompletion, Formatting, Hints, MarkOccurrences, ...]
 -> InlineHints = 0

Not all children in Windows2/Modes/editor/ marked with the position attribute:
[TreeDiffViewerTopComponent.wstcref, CoverageReportTopComponent.wstcref], but some are:
[DashboardDisplayer.wstcref]
 -> CoverageReportTopComponent = 0
 -> TreeDiffViewerTopComponent = 0

Not all children in Templates/Project/Maven2/ marked with the position attribute:
[settings.xml], but some are:
[JavaApp, org-netbeans-modules-maven-j2ee-ui-wizard-EEWizardIterator-createWebAppIterator, ...]
 -> settings.xml = 110

Not all children in / marked with the position attribute:
[org-netbeans-modules-lsp-client-bindings-CompletionProviderImpl.instance], but some are:
[org-netbeans-modules-editor-java-JavaCompletionProvider.instance, ...]
 -> CompletionProviderImpl = 1000

Not all children in / marked with the position attribute:
[org-netbeans-modules-spellchecker-completion-WordCompletion.instance], but some are:
[org-netbeans-modules-lsp-client-bindings-CompletionProviderImpl.instance@1000]
 -> WordCompletion = 600 (across several layer registrations)

Not all children in Editors/text/x-java/CodeGenerators/ marked with the position attribute:
[org-netbeans-modules-micronaut-db-MicronautDataEndpointGenerator$Factory.instance], but some are:
[org-netbeans-modules-j2ee-ejbcore-ui-logicalview-ejb-action-AddMethodActions$AddBusinessMethodCodeGenerator.instance@10, ...]
 -> MicronautDataEndpointGenerator$Factory = 800

Not all children in / marked with the position attribute:
[org-netbeans-modules-editor-java-JavaBracesMatcher.shadow], but some are:
[org-netbeans-modules-editor-bracesmatching-LegacyEssMatcher.instance@100, org-netbeans-modules-editor-bracesmatching-DefaultMatcher.instance@200]
 -> JavaBracesMatcher = 300

Not all children in / marked with the position attribute:
[org-netbeans-modules-xml-text-completion-XMLCompletionProvider.instance, org-netbeans-modules-xml-schema-completion-SchemaBasedCompletionProvider.instance], but some are:
[org-netbeans-modules-spellchecker-completion-WordCompletion.instance@600, ...]
 -> XMLCompletionProvider = 700
 -> SchemaBasedCompletionProvider = 800