Enable parsing of JSP 3.0 (Jakarta) based sources
JSP 2.X and 3.X could both be parsed by a 3.X compatible Parser. However
Jasper (the JSP implementation of Apache Tomcat) has ties to the API and
the parser thus has to match the target runtime environment.

To solve this problem both a JSP 2.X (javax-Namespace) and a 3.X parser
are built and switch based on the declared specification of the
project.

The servletjspapi module was enhanced to carry both the javax. and the
jakarta. based API classes. These don't clash - just many (all?)
implementations decided to not support deployment on the same classpath.

Closes: #8543