Diff: Fix partitial redrawing of change delimiter line after hover over icon
The LineNumbersActionsBar is used in the diff views it is the component
placed between the editor panes on both sides of the divider and the
divider. The component provides the "remove" and "add" line actions,
that allow users to align checked out state with the reference revision.

+--------------+-----------------------------------+
| Editor 1     |LNAB1 | DIV | LNBA2 |     Editor 2 |
|              |      |     |       |              |
+--------------+------+-----+-------+--------------+

Editor => Editor Pane
LNAB   => LineNumbersActionsBar
DIV    => Divider

The LineNumbersActionsBar itself is roughly (right side drawn below,
left side is inversed):

+----+----+
| LN | AI |
+----+----+

LN => LineNumber
AI => ActionIcon

The ActionIcon is redrawn when the mouse enters and exists its active
area. When that happens, the surrounding line is only partially drawn.

The problem is the usage of the clip rectangle. The two lines above and
below the action icon, that surround the individual diff entries were
drawn using the the values 0 and clip.width for the x coordinates of
the line. That works ok, when the full width is drawn, but fails in the
hover case.

This is the clip rectangle from a normal scroll:

    "java.awt.Rectangle[x=0,y=1546,width=48,height=102]"

And this from the hover:

    "java.awt.Rectangle[x=33,y=799,width=15,height=16]"

There is no need to redraw the whole area when just the action icon
changes state and in this case the drawn line is to short.

There is actually no reason to use the clip rectangle at this point.
The x coordinates are known. The line starts at the beginning (0) and
should be drawn over the whole width (barWidth). The drawing
architecture will clip correctly, we don't need to optimize here.