Improve editor tooltip mouse ignore area computation.
Some editor tooltips are interactive. To be able to enter the
tooltip with the mouse, a rectangle is computed which prevents the
tooltip to hide, even if the mouse leaves the component (which would
hide it instantly).

This improves this mechanism slightly by

 - adding an extra x-margin to the ignored area
 - offsetting the debugger tooltip to further move the buttons into
   the ignored area (right above the mouse cursor)

removed obsolete code in ToolTipAnnotation and minor cleanup in that
class (no other changes).